+++ player.h
@@ -666,7 +666,7 @@ class Player : public Creature, public Cylinder
 		void sendClosePrivate(uint16_t channelId) const
 			{if(client) client->sendClosePrivate(channelId);}
 		void sendIcons() const;
-		void sendMagicEffect(const Position& pos, uint8_t type) const
+		void sendMagicEffect(const Position& pos, uint16_t type) const
 			{if(client) client->sendMagicEffect(pos, type);}
 		void sendStats();
 		void sendSkills() const
